requirements.txt
flake8
numpy
pytest
pytest-cov
Pillow